fix: emulate "unique: false" in simple-transition demo
In v8, the default behavior of "useTransition" was to never reuse elements, which means new elements were created for every transition, even if a leaving element has the same key.

In v9, the "unique" option was removed and "unique: true" is the default behavior. To emulate "unique: false" in v9, you need to provide a new key for every transition, and then have some way to derive the correct element based off that key. In the "simple-transition" demo, that is pretty easy, since we just increment the "index" on every transition and use modulo to determine which element should be rendered.